+++ src/Phoenix.Models/Base/Dto/MetricDtoBase.cs
@@ -3,6 +3,6 @@ namespace Phoenix.Models.Base.Dto
    public abstract class MetricDtoBase : DtoBase
    {
       public bool IsActive { get; init; }
-      public uint Version { get; init; }
+      public short Version { get; init; }
    }
 }
